add alternative layout for related streams
