+++ app/src/main/res/layout/trending_row.xml
@@ -82,6 +82,8 @@
         android:id="@+id/textView_channel"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:ellipsize="end"
+        android:maxLines="2"
         android:paddingBottom="16dp"
         app:layout_constraintEnd_toEndOf="@+id/textView_title"
         app:layout_constraintStart_toStartOf="@+id/textView_title"
