+++ app/src/main/res/xml/player_settings.xml
@@ -55,6 +55,21 @@
 
     </PreferenceCategory>
 
+    <PreferenceCategory app:title="@string/appearance">
+
+        <SwitchPreferenceCompat
+            android:icon="@drawable/ic_caption"
+            app:defaultValue="true"
+            app:key="system_caption_style"
+            app:title="@string/system_caption_style" />
+
+        <Preference
+            android:icon="@drawable/ic_settings"
+            app:key="caption_settings"
+            app:title="@string/caption_settings" />
+
+    </PreferenceCategory>
+
     <PreferenceCategory app:title="@string/behavior">
 
         <SwitchPreferenceCompat
